pyproject.toml
[project]
name = "ai-employee-hackathon"
version = "0.1.0"
description = "Bronze Tier Constitutional FTE - Autonomous task orchestrator with safety guardrails"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.9.0",
    "python-dotenv>=1.0.0",
    "watchdog>=3.0.0",
    "pydantic>=2.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
